+++ src/components/Banner/index.tsx
@@ -1,39 +1,7 @@
 import React from "react";
 import { Subreddit } from "snoowrap";
-import styled from "styled-components";
 import SubredditIcon from "./SubredditIcon";
-
-type BannerWrapperProps = {
-  imgSrc: string;
-};
-
-const BannerWrapper = styled.div`
-  width: 100%;
-  height: 200px;
-  background-color: ${props => props.theme.primary};
-  background-image: ${(props: BannerWrapperProps) => `url(${props.imgSrc})`};
-  background-size: cover;
-  background-position: center;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: white;
-  padding: 40px;
-`;
-
-const Title = styled.h1`
-  display: flex;
-  align-items: center;
-
-  @media (max-width: 560px) {
-    font-size: 1.5em;
-  }
-
-  @media (max-width: 399px) {
-    flex-direction: column;
-    font-size: 1.2em;
-  }
-`;
+import { BannerWrapper, Title } from "./styles";
 
 type Props = {
   subreddit: Subreddit | null;
